+++ mc/1-8---1-12/src/main/java/dev/redstudio/redcore/utils/OptiNotFine.java
@@ -78,7 +78,7 @@ public final class OptiNotFine {
 
             return (boolean) shaderPackLoadedHandle.invoke();
         } catch (Throwable exception) {
-            RED_LOGGER.logFramedError("OptiNotFine", "Could not get OptiFine shaders status.", "If shaders are enabled things might break", exception.getMessage());
+            RED_LOGGER.framedError("OptiNotFine", "Could not get OptiFine shaders status.", "If shaders are enabled things might break", exception.getMessage());
             return false;
         }
     }
@@ -114,7 +114,7 @@ public final class OptiNotFine {
             if ((boolean) fastRenderGetterHandle.invoke(mc.gameSettings))
                 fastRenderSetterHandle.invoke(mc.gameSettings, false);
         } catch (Throwable exception) {
-            RED_LOGGER.logFramedError("OptiNotFine", "Could not disable OptiFine fast renderer", "Things will break", exception.getMessage());
+            RED_LOGGER.framedError("OptiNotFine", "Could not disable OptiFine fast renderer", "Things will break", exception.getMessage());
         }
     }