+++ Changelog.md
@@ -18,7 +18,7 @@ The method `RedLogger#printFramedError` has been renamed to `RedLogger#logFramed
   - `kebab-case`
   - `lower_snake_case`
   - `UPPER_SNAKE_CASE`
-- Added `RedLogger#logFramed` which allows logging lists in a readable manner (e.g., versioning information)
+- Added `RedLogger#framed` which allows logging lists in a readable manner (e.g., versioning information)
 - Added `DecimalUtils` a utility class which currently allows for truncating decimal numbers for human reading
 - Added empty private constructors to hide implicit public ones
 - Added `Vector3F` & `Vector2F` to the new experimental vector suite
@@ -26,7 +26,7 @@ The method `RedLogger#printFramedError` has been renamed to `RedLogger#logFramed
 ### Changed
 
 - Improved documentation for `RedLogger#logFramedError`
-- Renamed `RedLogger#printFramedError` to `RedLogger#logFramedError`
+- Renamed `RedLogger#printFramedError` to `RedLogger#framedError`
 
 ### Internal